mise en place de struct
